Fixed OOM when using lora on a 8GB GPU while saving model weights
Using float32 while saving the model weights causes a spike in vram usage while saving. This fixes that. Regular weights are already saved with float16 by default.

+++ dreambooth/train_dreambooth.py
@@ -889,7 +889,7 @@ def main(args: DreamboothConfig, memory_record, use_subdir, lora_model=None, lor
                                 args.pretrained_model_name_or_path,
                                 subfolder="unet",
                                 revision=args.revision,
-                                torch_dtype=torch.float32
+                                torch_dtype=torch.float16
                             ).to(accelerator.device)
                         else:
                             out_file = None
